Feat(testApp): add all Leaflet with plugin global version
for Rollup build, through rollup-plugin-inject to have the plugin expecting `global.L` being replaced by `import * as L from 'leaflet'`.
This works because the TestPlugin does not try to attach something directly on L namespace, but on its subnamespaces (i.e. L.Control and L.control). If it did, we would have needed to use rollup-plugin-replace (and possibly rollup-plugin-legacy) to replace the L.NewNameSpace by another local variable and export it.
Note: changed 'leaflet' import by 'leaflet2' to symlink to a local version of Leaflet that makes the version available as a build file, instead of having to import package.json.